% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_fan_count}
\alias{add_fan_count}
\title{Add variable fan_count to old existing data}
\usage{
add_fan_count(dir)
}
\arguments{
\item{dir}{A path to a directory containing data files generated with (older)
versions of this package.}
}
\description{
In version 3.2 of this package a new variable "fan_count" was added to
Facebook data. It records the fan count of pages when scraping posts. This
functions adds this variable (with NA) to all files in an folder (if it does
not exist already).
}
