% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_reactions}
\alias{add_reactions}
\title{Add variables for reactions to old existing data}
\usage{
add_reactions(dir, sort = TRUE)
}
\arguments{
\item{dir}{A path to a directory containing data files generated with this
package.}

\item{sort}{Should variables be sorted according to current package default?
Defaults to TRUE.}
}
\description{
If previous data was scraped without reactions, empty reaction variables are
added here if such variables to not exist already.
}
