% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_scraped_time}
\alias{add_scraped_time}
\title{Add variable scrape_time to old existing data}
\usage{
add_scraped_time(dir)
}
\arguments{
\item{dir}{A path to a directory containing data files generated with (older)
versions of this package.}
}
\description{
In version 3.1 of this package a new variable "scrape_time" was added. It
records the Sys.time() when the data was scraped. This functions adds this
variable (with NA) to all files in an folder.
}
