% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{current_facebook_sort}
\alias{current_facebook_sort}
\title{Current order of variables in Facebook data}
\usage{
current_facebook_sort()
}
\value{
The current list of variable names for sorting Facebook data.
}
\description{
Current order of variables in Facebook data
}
