% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_facebook_pages.R
\name{get_pageinfo}
\alias{get_pageinfo}
\title{Get additional information for pages}
\usage{
get_pageinfo(page, token, fields = "fan_count", api = NULL)
}
\arguments{
\item{page}{A page ID.}

\item{token}{Either a temporary access token created at
https://developers.facebook.com/tools/explorer or the OAuth token created
with fbOAuth.}

\item{fields}{The API fields that are requested. "fan_count" is the default.}

\item{api}{The API-version to use, e.g., "v2.8". NULL is the default.}
}
\value{
A named list where the page id and the additional fields are stored.
}
\description{
Get additional information for pages
}
