% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge rds files from different folders}
\usage{
merge_data(old_folder, new_folder, output_folder, id = "id",
  sort = "created_time", keep_newest = TRUE,
  ignore_scrape_time = FALSE)
}
\arguments{
\item{old_folder}{Folder to look for old rds files.}

\item{new_folder}{Folder to look for new rds files.}

\item{output_folder}{Folder to save merged rds files to.}

\item{id}{Unique id(s), only data where id is distinct is kept. Defaults to
"id" for Facebook data.}

\item{sort}{Merged data is sorted by this parameter. Defaults to
"created_time" for Facebook data.}

\item{keep_newest}{Logical, indicating which version of a duplicate text is
kept. If TRUE (default), the newest texts according to scrape date are
kept if ignore_scrape_time is not TRUE. Furthermore, texts from files in
the new_folder are preferred over those from old_folder. FALSE prefers
older data.}

\item{ignore_scrape_time}{Logical, indicating whether the scrape time should
be ignored for deciding which texts to keep. Defaults to FALSE. If TRUE,
only age only depends on where the file is stored (old or new folder).}
}
\value{
A data.frame with the results.
}
\description{
This function merges rds files of the same name from two different folders.
It is intended for usage with social media data collected through my update
functions in this package. Data from the file in old_folder is kept, data
from new_folder is added but only kept if rows with the same id do not
exist already.
}
