% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort_data}
\alias{sort_data}
\title{Sort a dataset}
\usage{
sort_data(data, sorted_variables)
}
\arguments{
\item{data}{A data.frame generated with this package.}

\item{sorted_variables}{A list of variable names according to which the
dataset is sorted. Defaults to the package default for Facebook data.}
}
\value{
The sorted dataset.
}
\description{
A social media dataset is ordered according to a list of variable names.
Unmentioned variables are attached at the end.
}
