% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort_dir}
\alias{sort_dir}
\title{Sort all data files from a directory}
\usage{
sort_dir(dir, sorted_variables)
}
\arguments{
\item{dir}{A path to a directory containing data files generated with this
package.}

\item{sorted_variables}{A list of variable names according to which the
dataset is sorted. Defaults to the package default for Facebook data.}
}
\description{
All datasets from one folder with rds files are ordered according to a list
of variable names. Unmentioned variables are attached at the end and the
original files are overwritten with the sorted datasets.
}
