% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort_file}
\alias{sort_file}
\title{Sort data stored in a rds file}
\usage{
sort_file(file, sorted_variables)
}
\arguments{
\item{file}{A path to an rds file generated with this package and containing
social media data.}

\item{sorted_variables}{A list of variable names according to which the
dataset is sorted. Defaults to the package default for Facebook data.}
}
\description{
A dataset stored in an rds file is ordered according to a list of variable
names. Unmentioned variables are attached at the end and the original file
is overwritten with the sorted dataset.
}
