% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pages.R
\name{update_pages}
\alias{update_pages}
\title{Update multiple Facebook pages}
\usage{
update_pages(pages = NULL, token = NULL, datadir = "./data",
  go_back = TRUE, n_posts = 100, feed = FALSE, reactions = TRUE)
}
\arguments{
\item{pages}{A named character vector. Names are the data file names,
characters are the page names that will be updated (see example).}

\item{token}{Either a temporary access token created at
https://developers.facebook.com/tools/explorer or the OAuth token created
with fbOAuth.}

\item{datadir}{A directory containing the stored data sets (as "name.rds").}

\item{go_back}{Go back in time, not only update newer posts, but also get
older ones. Default is TRUE.}

\item{n_posts}{The number of posts to scrape from each page, defaults to 100.}

\item{feed}{If TRUE, the function will also return posts from each page that
were made by others (not only the admin of the page).}

\item{reactions}{If TRUE, will add variables to the data frame with the total
count of reactions: love, haha, wow, sad, angry.}
}
\value{
TRUE for all pages the update went through.
}
\description{
This function updates multiple Facebook pages with \link{Rfacebook}'s
getPage. makes use of the \link{update_page} function in this package. For
each page in the provided character vector pages it looks for a previously
stored data.frame (stored in the datadir under the name provided in the named
vector pages with the file extension .rds). If there is one, it tries to
scrape newer (and older) posts than the ones already stored in the
data.frame. Otherwise it creats a new data.frame.
}
\examples{
# You have to authenticate with Facebook's API first. For more in this visit
# Rfacebooks documentation.
fb_oauth <- fbOAuth(app_id="123456789",
                    app_secret="1A2B3C4D",
                    extended_permissions = TRUE)

# A character vector is created
my_pages <- c(cnn = "cnn",
              bbc = "bbcnews")
# The two pages CNN and BBC news are updated and the results are stored in
# the user's home directory in the folder "temp". They are named cnn.rds and
# bbc.rds. If they do not already exist, both data sets are created.
update_pages(pages = my_pages, token = fb_outh, datadir = "~/temp")
}
