% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_twitter_data}
\alias{merge_twitter_data}
\title{Merge Twitter rds files from different folders}
\usage{
merge_twitter_data(old_folder, new_folder, output_folder,
  keep_newest = TRUE)
}
\arguments{
\item{old_folder}{Folder to look for old rds files.}

\item{new_folder}{Folder to look for new rds files.}

\item{output_folder}{Folder to save merged rds files to.}

\item{keep_newest}{Logical, indicating which version of a duplicate Tweet is
kept. If TRUE (default), the newest Tweets according to scrape date are
kept. Furthermore, Tweets from files in the new_folder are preferred over
those from old_folder. FALSE prefers older data.}
}
\value{
A data.frame with the results.
}
\description{
This function merges rds files of the same name from two different folders.
It calls merge_data with parameters for Facebook data collected through this
package's update function. Data from the file in old_folder is kept, data
from new_folder is added but only kept if rows with the same id do not
exist already.
}
